+++ main_test.go
@@ -6,8 +6,20 @@ import (
 )
 
 func TestYAMLToJSON(t *testing.T) {
+	indent = false
 	revert = false
 	expected := "{\"age\":30,\"name\":\"John\"}"
+	actual := string(processFile("test/simple.yaml"))
+
+	if strings.Compare(actual, expected) != 0 {
+		t.Errorf("Expected: %s, but was: %s", expected, actual)
+	}
+}
+
+func TestYAMLToJSONIndent(t *testing.T) {
+	indent = true
+	revert = false
+	expected := "{\n  \"age\": 30,\n  \"name\": \"John\"\n}"
 	actual := string(processFile("test/simple.yaml"))
 
 	if strings.Compare(actual, expected) != 0 {