+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+##### yaml2json 1.1.0 - (06 March 2017)
+
+* Indent JSON output
+
 ##### yaml2json 1.0.0 - (26 February 2017)
 
-    * First version
+* First version